requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.0
onnx
packaging==25.0
pandas==2.3.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
Pillow==10.2.0
torchvision==0.18.0
open-c2pa==0.2.6
